requirements.txt
# Django
Django>=5.1
django-htmx>=1.17.0
django-cotton>=1.0.0
whitenoise>=6.6.0

# Database
psycopg2-binary>=2.9.9

# AI & ML
pydantic>=2.5.0
pydantic-ai>=0.0.14
anthropic>=0.39.0

# Async
httpx>=0.27.0
aiohttp>=3.9.0

# News scraping
feedparser>=6.0.10
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8

# Payments
stripe>=7.0.0

# Utils
python-dotenv>=1.0.0
markdown>=3.5.0
Pillow>=10.0.0

# Dev
django-debug-toolbar>=4.2.0
pytest-django>=4.7.0
